Validation is more strict
Validation is more strict for the following states: AL, AK, AZ, AR, MO,
PR.

With these changes, not all DL that were previously valid will still be
valid.

Although we crosschecked with DL in production database, we're not
entirely convinced that there's a need to be more strict in these
validations.